+++ i2c/sensirion_i2c_hal.c
@@ -83,7 +83,7 @@ void sensirion_i2c_hal_free(void) {
  * @param count   number of bytes to read from I2C and store in the buffer
  * @returns 0 on success, error code otherwise
  */
-int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint16_t count) {
+int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint8_t count) {
     /* TODO:IMPLEMENT */
     return NOT_IMPLEMENTED_ERROR;
 }
@@ -100,7 +100,7 @@ int8_t sensirion_i2c_hal_read(uint8_t address, uint8_t* data, uint16_t count) {
  * @returns 0 on success, error code otherwise
  */
 int8_t sensirion_i2c_hal_write(uint8_t address, const uint8_t* data,
-                               uint16_t count) {
+                               uint8_t count) {
     /* TODO:IMPLEMENT */
     return NOT_IMPLEMENTED_ERROR;
 }
